fix: dryed the multiple implementations
